# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevDonna Outreach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
  contactCount: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  color?: string;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  color?: string | null;
}

export type ContactStatus = typeof ContactStatus[keyof typeof ContactStatus];


export const ContactStatus = {
  pending: 'pending',
  sent: 'sent',
  failed: 'failed',
} as const;

export interface Contact {
  id: number;
  companyName: string;
  /** @nullable */
  contactName?: string | null;
  email: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
  status: ContactStatus;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  companyName: string;
  contactName?: string;
  email: string;
  website?: string;
  industry?: string;
  categoryId?: number;
  city?: string;
  country?: string;
  phone?: string;
  notes?: string;
}

export interface ContactUpdate {
  /** @minLength 1 */
  companyName?: string;
  /** @nullable */
  contactName?: string | null;
  email?: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ContactList {
  contacts: Contact[];
  total: number;
  page: number;
  limit: number;
}

export type EmailRecordStatus = typeof EmailRecordStatus[keyof typeof EmailRecordStatus];


export const EmailRecordStatus = {
  sent: 'sent',
  failed: 'failed',
} as const;

export interface EmailRecord {
  id: number;
  campaignId: number;
  /** @nullable */
  campaignName?: string | null;
  contactId: number;
  contactEmail: string;
  /** @nullable */
  contactCompany?: string | null;
  status: EmailRecordStatus;
  /** @nullable */
  errorMessage?: string | null;
  sentAt: string;
}

export interface ContactDetail {
  contact: Contact;
  emails: EmailRecord[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export interface BulkDeleteInput {
  ids: number[];
}

export interface BulkDeleteResult {
  deleted: number;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  sending: 'sending',
  sent: 'sent',
  failed: 'failed',
} as const;

export interface Campaign {
  id: number;
  name: string;
  subject: string;
  body: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  status: CampaignStatus;
  sentCount: number;
  failedCount: number;
  totalCount: number;
  createdAt: string;
  /** @nullable */
  sentAt?: string | null;
}

export interface CampaignInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  body: string;
  categoryId: number;
}

export interface CampaignDetail {
  campaign: Campaign;
  emails: EmailRecord[];
}

export interface SendResult {
  sent: number;
  failed: number;
  total: number;
}

export interface EmailList {
  emails: EmailRecord[];
  total: number;
  page: number;
  limit: number;
}

export interface CategoryContactCount {
  categoryId: number;
  categoryName: string;
  count: number;
}

export interface DashboardStats {
  totalContacts: number;
  pendingContacts: number;
  totalCategories: number;
  totalEmailsSent: number;
  totalEmailsFailed: number;
  totalCampaigns: number;
  recentCampaigns: Campaign[];
  contactsByCategory: CategoryContactCount[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type ListContactsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
status?: ListContactsStatus;
/**
 * @nullable
 */
search?: string | null;
page?: number;
limit?: number;
};

export type ListContactsStatus = typeof ListContactsStatus[keyof typeof ListContactsStatus] | null;


export const ListContactsStatus = {
  pending: 'pending',
  sent: 'sent',
  failed: 'failed',
} as const;

export type ListEmailsParams = {
/**
 * @nullable
 */
campaignId?: number | null;
/**
 * @nullable
 */
status?: ListEmailsStatus;
page?: number;
limit?: number;
};

export type ListEmailsStatus = typeof ListEmailsStatus[keyof typeof ListEmailsStatus] | null;


export const ListEmailsStatus = {
  sent: 'sent',
  failed: 'failed',
} as const;