# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DevDonna Outreach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CONTACTS ─────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign with email stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Send campaign emails to all contacts in the selected category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"

  # ─── EMAILS ───────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: ["string", "null"]
        contactCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, contactCount, createdAt]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
      required: [name]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: ["string", "null"]

    # ─── CONTACTS ───────────────────────────────────────────────────
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sending, sent, failed]
        sentCount:
          type: integer
        failedCount:
          type: integer
        totalCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: ["string", "null"]
          format: date-time
      required: [id, name, subject, body, categoryId, status, sentCount, failedCount, totalCount, createdAt]
    CampaignInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        categoryId:
          type: integer
      required: [name, subject, body, categoryId]
    CampaignDetail:
      type: object
      properties:
        campaign:
          $ref: "#/components/schemas/Campaign"
        emails:
          type: array
          items:
            $ref: "#/components/schemas/EmailRecord"
      required: [campaign, emails]
    SendResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        total:
          type: integer
      required: [sent, failed, total]

    # ─── EMAILS ─────────────────────────────────────────────────────
    EmailRecord:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        campaignName:
          type: ["string", "null"]
        contactId:
          type: integer
        contactEmail:
          type: string
        contactCompany:
          type: ["string", "null"]
        status:
          type: string
          enum: [sent, failed]
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: string
          format: date-time
      required: [id, campaignId, contactId, contactEmail, status, sentAt]
...