# lib/api-zod/src/generated/types/emailRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevDonna Outreach API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailRecordStatus } from './emailRecordStatus';

export interface EmailRecord {
  id: number;
  campaignId: number;
  /** @nullable */
  campaignName?: string | null;
  contactId: number;
  contactEmail: string;
  /** @nullable */
  contactCompany?: string | null;
  status: EmailRecordStatus;
  /** @nullable */
  errorMessage?: string | null;
  sentAt: Date;
}