# lib/api-zod/src/generated/types/sendResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevDonna Outreach API
 * OpenAPI spec version: 0.1.0
 */

export interface SendResult {
  sent: number;
  failed: number;
  total: number;
}